# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/portal-openapi.yaml
openapi: 3.0.0
info:
    title: Gravitee.io Portal Rest API
    description: API dedicated to the devportal part of Gravitee
    termsOfService: terms
    contact:
        name: The Gravitee Team
        url: https://gravitee.io
        email: contact@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0
    version: "4.11.0-SNAPSHOT"
servers:
    - url: /portal/environments/{envId}
      description: The portal API for a given environment
      variables:
          envId:
              default: DEFAULT
              description: ID of the environment
...
security:
    - BasicAuth: []
    - CookieAuth: []
paths:
    /apis:
        get:
            tags:
                - Api
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
                - $ref: "#/components/parameters/apiContextPathQueryParam"
                - $ref: "#/components/parameters/apiLabelQueryParam"
                - $ref: "#/components/parameters/apiVersionQueryParam"
                - $ref: "#/components/parameters/apiNameQueryParam"
                - $ref: "#/components/parameters/apiCategoryQueryParam"
                - $ref: "#/components/parameters/apiFilterQueryParam"
                - $ref: "#/components/parameters/apiFilterExcludedQueryParam"
                - $ref: "#/components/parameters/apiPromotedQueryParam"
            summary: List APIs
            description: |
                List public APIs for anonymous requests.
                List all the APIs the current user is allowed to category for authenticated requests.

                The list can be filtered according to query parameters.

                By default the list is sorted by alphabetic order.
                If a **cat** query param is sent, a specific sorting can be applied. Please refer to **cat** description for detail.
            operationId: getApis
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List accessible APIs for current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApisResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /apis/categories:
        get:
            tags:
                - Api
            parameters:
                - $ref: "#/components/parameters/apiFilterQueryParam"
            summary: List Categories
            description: |
                For anonymous requests, list all categories of public APIs.
                For authenticated requests, list all categories of all the APIs the current user can access.

                The list is sorted by alphabetic order.
            operationId: listCategories
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List used categories in accessible APIs for current user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CategoriesResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /applications:
        get:
            tags:
                - Application
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
                - $ref: "#/components/parameters/forSubscriptionParam"
                - $ref: "#/components/parameters/applicationOrderQueryParam"
            summary: List all the applications accessible to authenticated user. Default order is by *name* ASC.
            description: |
                List all the applications accessible to authenticated user.

                User must have MANAGEMENT_APPLICATION[READ] and PORTAL_APPLICATION[READ] permission.
                For performance reasons, calling `/applications?size=-1` does not return the number of subscribers and the only information available in the owner of each application is:
                  - id
                  - displayName
                  - email
            operationId: getApplications
            responses:
                200:
                    description: List of Applications.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApplicationsResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
        post:
            tags:
                - Application
            requestBody:
                description: Use to create an application.
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ApplicationInput"
            summary: Create an application
            description: |
                Create an application.

                User must have MANAGEMENT_APPLICATION[CREATE] permission.
            operationId: createApplication
            responses:
                201:
                    description: Created application.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Application"
                400:
                    description: Bad Request if body is empty.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                403:
                    $ref: "#/components/responses/PermissionError"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /groups:
        get:
            tags:
                - Group
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
            summary: List groups
            description: |
                List all groups

                User must have the MANAGEMENT_GROUP[READ] permission.
            operationId: getGroups
            responses:
                200:
                    description: List of groups
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GroupsResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /categories:
        get:
            tags:
                - Portal
            parameters:
                - $ref: "#/components/parameters/pageNumberParam"
                - $ref: "#/components/parameters/pageSizeParam"
            summary: Get a Category list
            description: |
                Get all categories of the platform.
            operationId: getCategories
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            responses:
                200:
                    description: List of Categories
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CategoriesResponse"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
    /portal-navigation-items:
        get:
            tags:
                - Portal
            summary: Get the portal navigation items.
            operationId: getPortalNavigationItems
            description: |
                Get all the navigation items to be displayed in given area.
            security:
                - {}
                - BasicAuth: []
                - CookieAuth: []
            parameters:
                - name: area
                  in: query
                  required: true
                  description: The area of the navigation items (e.g., 'APPLICATION', 'API', etc.)
                  schema:
                      $ref: "#/components/schemas/PortalArea"
                - name: parentId
                  in: query
                  required: false
                  description: The parent ID to filter navigation items.
                  schema:
                      type: string
                - name: loadChildren
                  in: query
                  required: false
                  description: Whether to load child navigation items.
                  schema:
                      type: boolean
            responses:
                200:
                    $ref: "#/components/responses/PortalNavigationItemsSuccess"
                500:
                    $ref: "#/components/responses/InternalServerError"
...
components:
...
    parameters:
...
        pageNumberParam:
            name: page
            in: query
            required: false
            description: The page number for pagination.
            schema:
                type: integer
                default: 1
        pageSizeParam:
            name: size
            in: query
            required: false
            description: |
                The number of items per page for pagination.
                If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
                If the size is -1, the response contains all datas.
            schema:
                type: integer
                default: 10

        #Apis filtering
        apiContextPathQueryParam:
            name: context-path
            in: query
            required: false
            description: The context-path of an API.
            schema:
                type: string
        apiLabelQueryParam:
            name: label
            in: query
            required: false
            description: One of the labels of an API.
            schema:
                type: string
        apiVersionQueryParam:
            name: version
            in: query
            required: false
            description: The version of an API.
            schema:
                type: string
        apiNameQueryParam:
            name: name
            in: query
            required: false
            description: The name of an API.
            schema:
                type: string
        apiCategoryQueryParam:
            name: category
            in: query
            required: false
            description: One of the category linked with an API.
            schema:
                type: string
        apiFilterQueryParam:
            name: filter
            in: query
            required: false
            description: |
                Name of a FILTER. A filter is used to pre-select some APIs. It can be combined with others query params.
                  - FEATURED : top APIs. Sorted by specific order.
                  - MINE : APIs subscribed by the current user. Sorted by alphabetical order.
                  - STARRED: Rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
                  - TRENDINGS: Subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
            schema:
                $ref: "#/components/schemas/FilterApiQuery"
        apiFilterExcludedQueryParam:
            name: -filter
            in: query
            required: false
            description: |
                Name of a FILTER to exclude. Excluded filter has priority over filter. A filter is used to pre-select some APIs. It can be combined with others query params.
                  - FEATURED : not in top APIs. Sorted by specific order.
                  - MINE : APIs not subscribed by the current user. Sorted by alphabetical order.
                  - STARRED: not rated APIs. Sorted by Ratings DESC, nb of ratings DESC, alphabetic order.
                  - TRENDINGS: not subscribed APIs. Sorted by number of subscriptions DESC, alphabetic order.
            schema:
                $ref: "#/components/schemas/FilterApiQuery"
        apiPromotedQueryParam:
            name: promoted
            in: query
            required: false
            description: |
                Allow to filter on promoted APIs or not.
                  - If absent, no effect on the result
                  - If *false*, removes the promoted API of the list
                  - If *true*, returns only the promoted API of the list

                A promoted API is either the highlighted API of a category (if it exists) or the first API in the list.
            schema:
                type: boolean
...
        forSubscriptionParam:
            name: forSubscription
            in: query
            required: false
            description: If true, only return applications with APPLICATION_SUBSCRIPTION[CREATE] permission.
            schema:
                type: boolean
        applicationOrderQueryParam:
            name: order
            in: query
            required: false
            description: |
                By default, sort is ASC. If *field* starts with '-', the order sort is DESC.\
                Currently, only **name** and **nbSubscriptions** are supported.
            example: "order=-nbSubscriptions"
            schema:
                type: string

        #Pages filtering
...
    schemas:
...
        ErrorResponse:
            properties:
                errors:
                    description: List of errors
                    type: array
                    items:
                        $ref: "#/components/schemas/Error"
...
        ApisResponse:
            properties:
                data:
                    description: List of API.
                    type: array
                    items:
                        $ref: "#/components/schemas/Api"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        ApplicationsResponse:
            properties:
                data:
                    description: List of applications.
                    type: array
                    items:
                        $ref: "#/components/schemas/Application"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        GroupsResponse:
            properties:
                data:
                    description: List of groups.
                    type: array
                    items:
                        $ref: "#/components/schemas/Group"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        CategoriesResponse:
            properties:
                data:
                    description: List of categories.
                    type: array
                    items:
                        $ref: "#/components/schemas/Category"
                metadata:
                    $ref: "#/components/schemas/MetadataMap"
                links:
                    $ref: "#/components/schemas/Links"
...
        Error:
            properties:
                status:
                    description: Status HTTP of the error
                    type: string
                message:
                    description: Message of the error
                    type: string
                code:
                    description: Technical code of the error
                    type: string
                parameters:
                    description: Parameters
                    type: object
                    additionalProperties:
                        type: string
        Api:
            description: |
                Describes an API.
            required:
                - id
                - name
                - version
                - description
                - owner
            properties:
                id:
                    description: Unique identifier of an API.
                    type: string
                definitionVersion:
                    $ref: "#/components/schemas/DefinitionVersion"
                name:
                    description: Name of the API.
                    type: string
                version:
                    description: Version of the API.
                    type: string
                description:
                    description: Description of the API.
                    type: string
                type:
                    $ref: "#/components/schemas/ApiType"
                draft:
                    type: boolean
                    description: Whether or not the API is in draft.
                public:
                    type: boolean
                    description: Whether or not the API is public.
                running:
                    type: boolean
                    description: Whether or not the API is running.
                entrypoints:
                    description: List of all the available endpoints to call the API.
                    type: array
                    items:
                        type: string
                listener_type:
                    $ref: "#/components/schemas/ListenerType"
                labels:
                    description: List of labels linked to this API.
                    type: array
                    items:
                        type: string
                mcp:
                    $ref: "#/components/schemas/Mcp"
                owner:
                    $ref: "#/components/schemas/User"
                created_at:
                    description: Create date and time.
                    type: string
                    format: date-time
                updated_at:
                    description: Last update date and time.
                    type: string
                    format: date-time
                categories:
                    description: |
                        List of category keys this API belongs to.
                        Example: ["media", "health"]
                    type: array
                    items:
                        type: string
                rating_summary:
                    $ref: "#/components/schemas/RatingSummary"
                _links:
                    $ref: "#/components/schemas/ApiLinks"

                ###################
                # may be included #
                ###################
                pages:
                    description: |
                        Only returned with (*)/apis/{apiId}*.
                        Need *include* query param to contain 'pages'.

                        The documentation pages of this API. Same as (*)/apis/{apiId}/pages*.
                    type: array
                    items:
                        $ref: "#/components/schemas/Page"
                plans:
                    description: |
                        Only returned with (*)/apis/{apiId}*.
                        Need *include* query param to contain 'plans'.

                        The plans of this API. Same as (*)/apis/{apiId}/plans*.
                    type: array
                    items:
                        $ref: "#/components/schemas/Plan"
        DefinitionVersion:
            type: string
            description: API's gravitee definition version.
            example: V4
            enum:
                - V1
                - V2
                - V4
                - FEDERATED
        ListenerType:
            description: Listener of the API entrypoints.
            type: string
            example: HTTP
            enum:
                - HTTP
                - TCP
                - SUBSCRIPTION
                - KAFKA
        ApiType:
            description: Type of V4 APIs.
            type: string
            example: PROXY
            enum:
                - A2A_PROXY
                - LLM_PROXY
                - MCP_PROXY
                - PROXY
                - MESSAGE
                - NATIVE
...
        Group:
            required:
                - id
                - name
            properties:
                id:
                    description: Unique identifier of a group.
                    type: string
                name:
                    description: Name of the group.
                    type: string
...
        Application:
            properties:
                id:
                    description: Unique identifier of an application.
                    type: string
                name:
                    description: Name of the application.
                    type: string
                description:
                    description: Description of the application.
                    type: string
                domain:
                    description: Domain used by the application.
                    type: string
                applicationType:
                    description: Type of the application (Web, Mobile, ...).
                    type: string
                origin:
                    description: Application origin. e.g., MANAGEMENT, KUBERNETES, INTEGRATION.
                    type: string
                hasClientId:
                    description: True if the application has client id.
                    type: boolean
                owner:
                    $ref: "#/components/schemas/User"
                created_at:
                    description: Creation date and time of the application.
                    type: string
                    format: date-time
                updated_at:
                    description: Last update date and time of the application.
                    type: string
                    format: date-time
                groups:
                    description: Array of groups associated to the application.
                    type: array
                    items:
                        $ref: "#/components/schemas/Group"
                picture:
                    description: |
                        Picture of the application. This attribute is only used to update a picture.\
                        To get the application picture, use /application/{applicationId}/picture.
                    type: string
                    writeOnly: true
                background:
                    description: |
                        Background of the application. This attribute is only used to update a picture.\
                        To get the application picture, use /application/{applicationId}/background.
                    type: string
                    writeOnly: true
                api_key_mode:
                    $ref: "#/components/schemas/ApiKeyModeEnum"
                settings:
                    $ref: "#/components/schemas/ApplicationSettings"
                _links:
                    $ref: "#/components/schemas/ApplicationLinks"
...
        User:
            properties:
                id:
                    description: Unique identifier of a user.
                    type: string
                reference:
                    description: Unique reference if user comes from external source. Use for search only.
                    type: string
                first_name:
                    type: string
                last_name:
                    type: string
                display_name:
                    type: string
                email:
                    type: string
                editable_profile:
                    description: True if the user can edit the MyAccount information
                    type: boolean
                    default: false
                permissions:
                    description: Map of Permissions.
                    type: object
                    properties:
                        APPLICATION:
                            type: array
                            items:
                                type: string
                        USER:
                            type: array
                            items:
                                type: string
                customFields:
                    type: object
                    description: Values for CustomUserFields
                    additionalProperties: true
                config:
                    $ref: "#/components/schemas/UserConfig"
                _links:
                    $ref: "#/components/schemas/UserLinks"
...
        Category:
            properties:
                id:
                    description: Unique identifier of a category.
                    type: string
                name:
                    type: string
                description:
                    type: string
                order:
                    type: integer
                page:
                    description: id of the page used for category documentation.
                    type: string
                total_apis:
                    type: integer
                    format: int64
                _links:
                    $ref: "#/components/schemas/CategoryLinks"
...
        ApplicationInput:
            properties:
                name:
                    type: string
                description:
                    type: string
                domain:
                    type: string
                picture:
                    type: string
                groups:
                    type: array
                    items:
                        type: string
                settings:
                    $ref: "#/components/schemas/ApplicationSettings"
                background:
                    type: string
                api_key_mode:
                    $ref: "#/components/schemas/ApiKeyModeEnum"
...
        ApplicationSettings:
            properties:
                app:
                    $ref: "#/components/schemas/SimpleApplicationSettings"
                oauth:
                    $ref: "#/components/schemas/OAuthClientSettings"
                tls:
                    $ref: "#/components/schemas/TlsClientSettings"
...
        ApiKeyModeEnum:
            description: |
                The API Key mode to use for this application.
                  - The `SHARED` API Key mode allows consumer to use the same API Key across all the subscriptions
                  - The `EXCLUSIVE` API Key mode will result to a new API Key being generated for each subscription
                  - The `UNSPECIFIED` API Key mode is a marker value informing that no choice as been made yet regarding
                    the API Key mode to use for the application.
            type: string
            enum:
                - SHARED
                - EXCLUSIVE
                - UNSPECIFIED
...
        MetadataMap:
            description: Map of Map of Object
            type: object
            additionalProperties:
                type: object
                additionalProperties:
                    type: object
...
        FilterApiQuery:
            type: string
            enum:
                - ALL
                - FEATURED
                - MINE
                - STARRED
                - TRENDINGS
...
        Links:
            properties:
                self:
                    type: string
                first:
                    type: string
                last:
                    type: string
                prev:
                    type: string
                next:
                    type: string
...
        ApplicationLinks:
            properties:
                self:
                    type: string
                members:
                    type: string
                notifications:
                    type: string
                picture:
                    type: string
                background:
                    type: string
...
        PortalArea:
            type: string
            description: The portal area (used by portal navigation items)
            enum: [HOMEPAGE, TOP_NAVBAR]
...
        PortalNavigationItem:
            oneOf:
                - $ref: "#/components/schemas/PortalNavigationPage"
                - $ref: "#/components/schemas/PortalNavigationFolder"
                - $ref: "#/components/schemas/PortalNavigationLink"
            discriminator:
                propertyName: type
                mapping:
                    PAGE: "#/components/schemas/PortalNavigationPage"
                    FOLDER: "#/components/schemas/PortalNavigationFolder"
                    LINK: "#/components/schemas/PortalNavigationLink"
...
    responses:
        InternalServerError:
            description: Internal Server Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
        PermissionError:
            description: Insufficient rights to access this resource
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
...
        PortalNavigationItemsSuccess:
            description: List of portal navigation items
            content:
                application/json:
                    schema:
                        type: array
                        items:
                            $ref: "#/components/schemas/PortalNavigationItem"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM